source/mars_consumer/requirements.txt
confluent-kafka==2.3.0
requests==2.31.0
redis==5.0.1
psycopg2-binary==2.9.9
fastapi==0.104.1
uvicorn==0.24.0